software_module/requirements.txt
numpy
pandas
numba
BioPython
scipy
logomaker
matplotlib
statsmodels